feat: promote pytorch custom reducers from experimental
