+++ harness/determined/pytorch/__init__.py
@@ -11,7 +11,13 @@ from determined.pytorch._data import (
 )
 from determined.pytorch._callback import PyTorchCallback
 from determined.pytorch._lr_scheduler import LRScheduler
-from determined.pytorch._reducer import MetricReducer, _SimpleReducer, Reducer, _reduce_metrics
+from determined.pytorch._reducer import (
+    MetricReducer,
+    _PyTorchReducerContext,
+    _SimpleReducer,
+    Reducer,
+    _reduce_metrics,
+)
 from determined.pytorch._experimental import PyTorchExperimentalContext
 from determined.pytorch._pytorch_context import PyTorchTrialContext
 from determined.pytorch._pytorch_trial import PyTorchTrial, PyTorchTrialController, reset_parameters
