+++ harness/determined/pytorch/_pytorch_context.py
@@ -29,7 +29,7 @@ except ImportError:
     pass
 
 
-class PyTorchTrialContext(det.TrialContext):
+class PyTorchTrialContext(det.TrialContext, pytorch._PyTorchReducerContext):
     """Contains runtime information for any Determined workflow that uses the ``PyTorch`` API.
 
     With this class, users can do the following things:
@@ -47,7 +47,8 @@ class PyTorchTrialContext(det.TrialContext):
     """
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
+        det.TrialContext.__init__(self, *args, **kwargs)
+        pytorch._PyTorchReducerContext.__init__(self)
 
         self._init_device()
 
@@ -75,6 +76,7 @@ class PyTorchTrialContext(det.TrialContext):
         self._current_batch_idx = None  # type: Optional[int]
 
         self.experimental = pytorch.PyTorchExperimentalContext(self)
+        self._reducers = pytorch._PyTorchReducerContext()
 
     def autocast_forward_pass(self, to_wrap: torch.nn.Module) -> torch.nn.Module:
         # First, ensure the forward pass is wrapped in an autocast context:
